requirements.txt
asgiref==3.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.54.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.10
idna==3.4
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.6.3
urllib3==2.0.3
